Move to_act into lib.rs
This is the bulk of the program so it made sense to me for it to be here
